% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_list.R
\name{AIC.list}
\alias{AIC.list}
\title{List method for An Information Criterion}
\usage{
\method{AIC}{list}(object, ..., k = 2)
}
\arguments{
\item{object}{list of fitted model objects for which logLik method exists}

\item{...}{ignored with a warning}

\item{k}{numeric, the penalty per parameter to be used; default k = 2 is the classical AIC}
}
\value{
A `data.frame` with rows corresponding to the members of object,
and columns for the number of parameters in the model and the IC.
}
\description{
List method for An Information Criterion
}
