% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_gam.R
\name{augment.gam}
\alias{augment.gam}
\title{An alternative augment approach}
\usage{
\method{augment}{gam}(model)
}
\arguments{
\item{model}{}
}
\value{
a dataframe with the input data, fitted values, and regression diagnostics
}
\description{
An alternative augment approach
}
\examples{
x <- mgcv::gam(Sepal.Width ~ s(Petal.Length), data = iris)
augment.other(x)

dat <- mgcv::gamSim(1,n=400,dist="binary",scale=.33)
lr.fit <- mgcv::gam(y~s(x0)+s(x1)+s(x2)+s(x3),family=binomial,
              data=dat,method="REML")

}
