% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_gam.R
\name{augment.other}
\alias{augment.other}
\title{An alternative augment approach}
\usage{
augment.other(model)
}
\arguments{
\item{model}{}
}
\value{
a dataframe with the input data, fitted values, and regression diagnostics
}
\description{
An alternative augment approach
}
\examples{
x <- mgcv::gam(Sepal.Width ~ s(Petal.Length), data = iris)
augment.other(x)

}
