% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_gam.R
\name{augment_gam}
\alias{augment_gam}
\title{Augment GAM models akin to broom::augments. .}
\usage{
augment_gam(x)
}
\arguments{
\item{x}{a fitted GAM object.}

\item{...}{not used at present}
}
\value{
NULL called for its side effect, making a plot
}
\description{
This function replicates broom::augment() for GAMs.
}
\examples{
x <- mgcv::gam(Sepal.Width ~ s(Petal.Length), data = iris)
augment_gam(x)

}
