% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gam.R
\name{check_gam}
\alias{check_gam}
\title{Plot residuals of a fitted generalised additive model.}
\usage{
check_gam(x, ...)
}
\arguments{
\item{x}{a fitted GAM object that has broom tidiers}

\item{...}{not used at present}
}
\value{
NULL called for its side effect, making a plot
}
\description{
This function replicates plot.lm() functionality with ggplot
}
\examples{
x <- mgcv::gam(Sepal.Width ~ s(Petal.Length), data = iris)
check_gam(x)

}
