% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nonlinear.R
\name{check_nonlinear}
\alias{check_nonlinear}
\title{Plot model parameters against residuals of a fitted model}
\usage{
check_nonlinear(x, ...)
}
\arguments{
\item{x}{a fitted model object}

\item{...}{not used at present}
}
\value{
NULL called for its side effect, making a plot
}
\description{
Plot model parameters against residuals of a fitted model
}
\examples{
lm1 <- lm(Sepal.Width ~ Petal.Length, data = iris)
check_nonlinear(lm1) # some indication of issues
lm2 <- lm(Sepal.Width ~ Petal.Length*Petal.Width, data = iris)
check_nonlinear(lm2)
# Also works with categorical variables.
lm3 <- lm(Sepal.Width ~ Petal.Length*Petal.Width*Species, data = iris)
check_nonlinear(lm3) # fixes the problems with lm1 and lm2
}
