% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRES803.R
\docType{package}
\name{NRES803}
\alias{NRES803}
\title{Data and functions for NRES 803 Ecological Statistics}
\description{
This package combines all the datasets necessary for labs and homeworks in NRES803.
I will also add useful functions that we create during the semester.
}
