% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRES803.R
\docType{package}
\name{NRES803}
\alias{NRES803}
\title{Functions for NRES 803 Ecological Statistics}
\description{
This package contains useful functions for students in NRES/STAT 803 Ecological Statistics at the University of Nebraska-Lincoln.
}
